# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/add-book": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Some Name"
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "Some Name"
                                            },
                                            "publish_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2016-05-26"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": "12"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Some Name"
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "Some Name"
                                        },
                                        "publish_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "Some Date"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "12"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-11-08T06:56:27.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-11-08T06:56:27.000000Z"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": "122"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/books": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "all books",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "122"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Some Name"
                                                    },
                                                    "author": {
                                                        "type": "string",
                                                        "example": "Some Name"
                                                    },
                                                    "publish_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "Some Date"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": "12"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2023-11-08T06:56:27.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2023-11-08T06:56:27.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/book/{book}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "show book",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "ID Book",
                        "example": 28
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "28"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Some Name"
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "Some Name"
                                        },
                                        "publish_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "Some Date"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "12"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-11-08T06:56:27.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-11-08T06:56:27.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update-book/{book}": {
            "put": {
                "tags": [
                    "Book"
                ],
                "summary": "update book",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "ID Book",
                        "example": 28
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Some Name"
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "Some Name"
                                            },
                                            "publish_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2016-05-26"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book Updated."
                                        },
                                        "Book": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "28"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some Name"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Some Name"
                                                },
                                                "publish_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2016-05-26"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-book/{book}": {
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "delete book",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "ID Book",
                        "example": 28
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book Deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}